+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 
 ## Unreleased
+### Added
+- `region_continent` field to `Facility`
 
 
 ## 2.9.0